requirements.txt
attrs==23.1.0
aws-cdk-lib==2.117.0
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
boto3==1.28.49
boto3-stubs==1.34.29
botocore==1.31.49
botocore-stubs==1.34.29
cattrs==23.2.3
constructs==10.3.0
exceptiongroup==1.2.0
gitdb==4.0.11
GitPython==3.1.40
importlib-resources==6.1.1
jmespath==1.0.1
jsii==1.93.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.6.2
six==1.16.0
smmap==5.0.1
typeguard==2.13.3
types-awscrt==0.20.3
types-s3transfer==0.10.0
typing_extensions==4.9.0
urllib3==1.26.16
